refactor Input and TextArea to a new Inputs module
