+++ plop-templates/Component/Component.tsx.hbs
@@ -7,7 +7,7 @@ interface Props {
 
 const {{pascalCase name}}: React.FC<Props> = ({ children }) => {
   return (
-    <div className={styles.{{camelCase name}}\}>
+    <div className={styles.{{camelCase name}}}>
       {{pascalCase name}}
       {children}
     </div>
