+++ src/components/Inputs/Input/test/ProjectSearch.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import ProjectSearch from "../ProjectSearch";
+import Input from "../Input";
 
 describe("ProjectSearch", () => {
   it("renders without error", () => {});
