+++ src/components/Inputs/index.ts
@@ -1,0 +1,4 @@
+import Input from "./Input";
+import TextArea from "./TextArea";
+
+export { Input, TextArea };
