--- src/components/ProjectSearch/index.ts
@@ -1,3 +1,0 @@
-import ProjectSearch from './ProjectSearch';
-
-export default ProjectSearch;
