+++ src/components/index.ts
@@ -1,9 +1,9 @@
 /* PLOP_INJECT_IMPORT */
-import TextArea from "./TextArea";
+import { Input, TextArea } from "./Inputs";
 import ButtonGroup from "./ButtonGroup";
 import TimerDisplay from "./TimerDisplay";
 import App from "./App";
-import ProjectSearch from "./ProjectSearch";
+import ProjectSearch from "./Inputs/Input";
 import Card from "./Card";
 import Brand from "./Brand";
 import Header from "./Header";
@@ -11,14 +11,15 @@ import { PrimaryButton, SecondaryButton } from "./Button";
 
 export {
   /* PLOP_INJECT_EXPORT */
-	TextArea,
-	ButtonGroup,
-	TimerDisplay,
-	App,
-	ProjectSearch,
-	Card,
-	Brand,
-	Header,
+  Input,
+  TextArea,
+  ButtonGroup,
+  TimerDisplay,
+  App,
+  ProjectSearch,
+  Card,
+  Brand,
+  Header,
   PrimaryButton,
   SecondaryButton,
 };
